♻️ refactor: extract invitation button component on header
