+++ src/components/header/Header.tsx
@@ -6,10 +6,8 @@ import Button from '../Button';
 import Title from './Title';
 import useDashboardStore from '@/store/dashboardStore';
 import UserSection from './UserSection';
-import useModalStore from '@/store/modalStore';
-import InvitationModal from '@/app/(with-header-sidebar)/dashboard/[id]/edit/_components/InvitationModal';
-import useInvitation from '@/app/(with-header-sidebar)/dashboard/[id]/edit/_hooks/useInvitation';
 import styles from './Header.module.css';
+import InvitationButton from './InvitationButton';
 
 interface HeaderProps {
   component?: React.ComponentType;
@@ -18,13 +16,6 @@ interface HeaderProps {
 export default function Header({ component: Component }: HeaderProps) {
   const router = useRouter();
   const dashboard = useDashboardStore((state) => state.dashboard);
-
-  const { openModal } = useModalStore();
-  const { handleInvite } = useInvitation(dashboard?.id.toString() || null);
-
-  const handleOpenInvitationModal = () => {
-    openModal(<InvitationModal handleInvite={handleInvite} />);
-  };
 
   const handleSettingsClick = () => {
     router.push(`/dashboard/${dashboard?.id}/edit`);
@@ -46,16 +37,7 @@ export default function Header({ component: Component }: HeaderProps) {
             관리
           </Button>
         )}
-        <Button className={styles.button} onClick={handleOpenInvitationModal}>
-          <Image
-            src="/icons/add_box.svg"
-            alt="초대하기"
-            width={20}
-            height={20}
-            className={styles.icon}
-          />
-          초대하기
-        </Button>
+        {dashboard && <InvitationButton dashboardId={dashboard.id} />}
       </div>
       {Component && (
         <div>
