+++ src/components/header/InvitationButton.module.css
@@ -1,0 +1,17 @@
+.button.button {
+  padding: 6px 12px;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  background-color: transparent;
+  border: 1px solid var(--gray-300);
+  border-radius: 6px;
+  color: var(--gray-500);
+  font-size: 14px;
+  font-weight: 500;
+  transition: background-color 0.3s ease;
+}
+
+.button.button:hover {
+  background-color: var(--violet-light);
+}
